move lucene query related functions to KeywordSource
